Modify tid from uint32_t to int

+++ js_environment/frameworks/js_environment/src/js_environment.cpp
@@ -178,7 +178,7 @@ bool JsEnvironment::StartDebugger(
         JSENV_LOG_E("Abnormal parsing of tid results.");
         return false;
     }
-    debugMode_ = panda::JSNApi::StartDebuggerForSocketPair(static_cast<uint32_t>(identifierId), socketFd);
+    debugMode_ = panda::JSNApi::StartDebuggerForSocketPair(identifierId, socketFd);
     return debugMode_;
 }
 
@@ -199,7 +199,7 @@ void JsEnvironment::StopDebugger(std::string& option)
         JSENV_LOG_E("Abnormal parsing of tid results.");
         return;
     }
-    panda::JSNApi::StopDebugger(static_cast<uint32_t>(identifierId));
+    panda::JSNApi::StopDebugger(identifierId);
 }
 
 void JsEnvironment::InitConsoleModule()
@@ -250,7 +250,7 @@ bool JsEnvironment::LoadScript(const std::string& path, uint8_t* buffer, size_t 
 }
 
 void JsEnvironment::StartProfiler(const char* libraryPath, uint32_t instanceId, PROFILERTYPE profiler,
-    int32_t interval, uint32_t tid, bool isDebugApp)
+    int32_t interval, int tid, bool isDebugApp)
 {
     if (vm_ == nullptr) {
         JSENV_LOG_E("Invalid vm.");
@@ -315,7 +315,7 @@ void JsEnvironment::SetDeviceDisconnectCallback(const std::function<bool()> &cb)
 }
 
 void JsEnvironment::NotifyDebugMode(
-    uint32_t tid, const char* libraryPath, uint32_t instanceId, bool debug, bool debugMode)
+    int tid, const char* libraryPath, uint32_t instanceId, bool debug, bool debugMode)
 {
     if (vm_ == nullptr) {
         JSENV_LOG_E("Invalid vm.");
